bug fixing.

+++ app/src/main/java/com/teamx/equiz/data/models/orderDetailData/Orders.kt
@@ -12,6 +12,6 @@ data class Orders(
     val orderStatus: String,
     val paymentIntent: String,
     val shippingInfo: ShippingInfo,
-    val totalPoints: Int,
+    val totalPoints: Double,
     val user: User
 )
